Renamed ScopedLoggingLevelSwitch and its interface.
