+++ src/Serilog/Serilog.csproj
@@ -47,7 +47,7 @@
 
   <ItemGroup Condition=" '$(TargetFramework)' == 'net45' OR '$(TargetFramework)' == 'net46' OR '$(TargetFramework)' == 'netstandard1.0' OR '$(TargetFramework)' == 'netstandard1.3' OR '$(TargetFramework)' == 'netstandard2.0' ">
     <Compile Remove="Core\OpenSwitchLogger.cs" />
-    <Compile Remove="Core\LoggingLevelOverrider.cs" />
+    <Compile Remove="Core\ScopedLoggingLevelSwitch.cs" />
     <Compile Remove="OpenSwitchLoggerExtensions.cs" />
   </ItemGroup>
